refactor: move uploadthing token util and type to shared file
